+++ docs/html/navtreeindex6.js
@@ -1,5 +1,8 @@
 var NAVTREEINDEX6 =
 {
+"namespacemembers_g.html":[15,1,0,6],
+"namespacemembers_h.html":[15,1,0,7],
+"namespacemembers_i.html":[15,1,0,8],
 "namespacemembers_k.html":[15,1,0,9],
 "namespacemembers_l.html":[15,1,0,10],
 "namespacemembers_m.html":[15,1,0,11],
@@ -246,8 +249,5 @@ var NAVTREEINDEX6 =
 "struct_session_state.html#aafdada6c72b108a7b369426eef84666c":[16,0,24,18],
 "struct_session_state.html#ab0391031ffb3a039c6db4e17f06ae93c":[16,0,24,14],
 "struct_session_state.html#ab6ffa48d10fbd8927ca80a96a195a086":[16,0,24,2],
-"struct_session_state.html#ab8e72d0a538b87295dfdf147c7a58e90":[16,0,24,5],
-"struct_session_state.html#aca10590c2b3dbbe1c2aa6d1e46d0e6b6":[16,0,24,9],
-"struct_session_state.html#ae0bcfaae63262ccee6e4aec733e6f7e1":[16,0,24,6],
-"struct_session_state.html#af3394279a59294a89ada8607297649be":[16,0,24,3]
+"struct_session_state.html#ab8e72d0a538b87295dfdf147c7a58e90":[16,0,24,5]
 };