+++ docs/html/search/all_7.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['7_0',['Days 4-7',['../md_docs_2v11-soak-plan.html#autotoc_md334',1,'']]],
+  ['7_0',['Days 4-7',['../md_docs_2v11-soak-plan.html#autotoc_md335',1,'']]],
   ['7_2000_202026_2004_2021_1',['7.00 - 2026-04-21',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md70',1,'']]],
-  ['7_20connectivity_20architecture_2',['7. Connectivity Architecture',['../md_docs_2architecture-overview.html#autotoc_md142',1,'']]],
-  ['7_20pirsensor_20cpp_20medium_20priority_3',['7. PIRSensor.cpp (MEDIUM PRIORITY)',['../md_docs_2releases_2v3_823-implementation-status.html#autotoc_md301',1,'']]]
+  ['7_20connectivity_20architecture_2',['7. Connectivity Architecture',['../md_docs_2architecture-overview.html#autotoc_md143',1,'']]],
+  ['7_20pirsensor_20cpp_20medium_20priority_3',['7. PIRSensor.cpp (MEDIUM PRIORITY)',['../md_docs_2releases_2v3_823-implementation-status.html#autotoc_md302',1,'']]]
 ];