+++ docs/html/search/pages_13.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['key_20diagnostic_20breakthrough_0',['Key Diagnostic Breakthrough',['../md_docs_2postmortems_22026-06-pmic-charging-root-cause.html#autotoc_md264',1,'']]],
-  ['known_20failure_20signature_1',['Known Failure Signature',['../md_docs_2postmortems_22026-06-pmic-charging-root-cause.html#autotoc_md268',1,'']]],
+  ['key_20diagnostic_20breakthrough_0',['Key Diagnostic Breakthrough',['../md_docs_2postmortems_22026-06-pmic-charging-root-cause.html#autotoc_md265',1,'']]],
+  ['known_20failure_20signature_1',['Known Failure Signature',['../md_docs_2postmortems_22026-06-pmic-charging-root-cause.html#autotoc_md269',1,'']]],
   ['known_20limitations_2',['Known Limitations',['../md__c_h_a_n_g_e_l_o_g.html#autotoc_md57',1,'']]]
 ];